Master

+++ migrations/versions/17dcb75f3fe_changed_the_type_of_the_column_last_.py
@@ -10,20 +10,25 @@ Create Date: 2015-03-10 14:20:53.676344
 revision = '17dcb75f3fe'
 down_revision = 'cde34831ea'
 
+from datetime import datetime
+import conf
 from alembic import op
 import sqlalchemy as sa
 
-from datetime import datetime
 
 def upgrade():
     unix_start = datetime(1970, 1, 1)
-    op.drop_column('feed', 'last_modified')
-    op.add_column('feed', sa.Column('last_modified', sa.String(),
-            nullable=True, default=unix_start, server_default=str(unix_start)))
+    if 'sqlite' not in conf.SQLALCHEMY_DATABASE_URI:
+        op.drop_column('feed', 'last_modified')
+        op.add_column('feed', sa.Column('last_modified', sa.String(),
+                      nullable=True, default=unix_start,
+                      server_default=str(unix_start)))
 
 
 def downgrade():
     unix_start = datetime(1970, 1, 1)
-    op.drop_column('feed', 'last_modified')
-    op.add_column('feed', sa.Column('last_modified', sa.DateTime(),
-            nullable=True, default=unix_start, server_default=unix_start))
+    if 'sqlite' not in conf.SQLALCHEMY_DATABASE_URI:
+        op.drop_column('feed', 'last_modified')
+        op.add_column('feed', sa.Column('last_modified', sa.DateTime(),
+                      nullable=True, default=unix_start,
+                      server_default=unix_start))
